Try scaling number of GMS instances a bit (#29)
### Notify
cc stripe-private-oss-forks/schema-and-query
r? divyamanohar || szhou || tgooch


### Summary
Scaling out the GMS service a bit. We're currently at 1 instance in qa and 3 in prod. As we dial up traffic, we are seeing ingest calls occasionally taking 1-3s. Trying to dig a bit into bottlenecks but indicators so far seem to suggest that RDS and ES seem ok. There were a couple of suggestions on DataHub slack to scale up GMS so trying that to see if the bottleneck is there (also reaching out to the Datahub folks to get a sense for which metrics we can explore on this)

### Motivation

### Testing

### Rollout/monitoring/revert plan
Safe to rollback.

(Squashed by Merge Queue - Original PR: https://git.corp.stripe.com/stripe-private-oss-forks/datahub/pull/29)